+++ airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/dataflow/input/DataFlowPipelineInputFlow.kt
@@ -4,13 +4,15 @@
 
 package io.airbyte.cdk.load.dataflow.input
 
-import io.airbyte.cdk.load.dataflow.DataFlowStageIO
+import io.airbyte.cdk.load.dataflow.finalization.StreamCompletionTracker
+import io.airbyte.cdk.load.dataflow.pipeline.DataFlowStageIO
 import io.airbyte.cdk.load.dataflow.state.StateKeyClient
 import io.airbyte.cdk.load.dataflow.state.StateStore
 import io.airbyte.cdk.load.message.CheckpointMessage
 import io.airbyte.cdk.load.message.DestinationMessage
 import io.airbyte.cdk.load.message.DestinationRecord
-import jakarta.inject.Singleton
+import io.airbyte.cdk.load.message.DestinationRecordStreamComplete
+import io.github.oshai.kotlinlogging.KotlinLogging
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.FlowCollector
 
@@ -20,12 +22,14 @@ import kotlinx.coroutines.flow.FlowCollector
  * Adds state ids to the input, handling the serial case where we infer the state id from a global
  * counter.
  */
-@Singleton
 class DataFlowPipelineInputFlow(
     private val inputFlow: Flow<DestinationMessage>,
     private val stateStore: StateStore,
     private val stateKeyClient: StateKeyClient,
+    private val completionTracker: StreamCompletionTracker,
 ) : Flow<DataFlowStageIO> {
+    val log = KotlinLogging.logger {}
+
     override suspend fun collect(
         collector: FlowCollector<DataFlowStageIO>,
     ) {
@@ -41,8 +45,11 @@ class DataFlowPipelineInputFlow(
                         )
                     collector.emit(io)
                 }
+                is DestinationRecordStreamComplete -> completionTracker.accept(it)
                 else -> Unit
             }
         }
+
+        log.info { "Finished routing input." }
     }
 }