+++ airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/dataflow/stages/StateStage.kt
@@ -4,9 +4,9 @@
 
 package io.airbyte.cdk.load.dataflow.stages
 
-import io.airbyte.cdk.load.dataflow.DataFlowStage
-import io.airbyte.cdk.load.dataflow.DataFlowStageIO
-import io.airbyte.cdk.load.dataflow.state.StateStore
+import io.airbyte.cdk.load.dataflow.pipeline.DataFlowStage
+import io.airbyte.cdk.load.dataflow.pipeline.DataFlowStageIO
+import io.airbyte.cdk.load.dataflow.state.StateHistogramStore
 import io.github.oshai.kotlinlogging.KotlinLogging
 import jakarta.inject.Named
 import jakarta.inject.Singleton
@@ -14,14 +14,16 @@ import jakarta.inject.Singleton
 @Named("state")
 @Singleton
 class StateStage(
-    val stateStore: StateStore,
+    private val stateHistogramStore: StateHistogramStore,
 ) : DataFlowStage {
     private val log = KotlinLogging.logger {}
 
     override suspend fun apply(input: DataFlowStageIO): DataFlowStageIO {
-        val stateUpdates = input.partitionHistogram!!
+        val countUpdates = input.partitionCountsHistogram!!
+        val byteUpdates = input.partitionBytesHistogram!!
 
-        stateStore.acceptFlushedCounts(stateUpdates)
+        stateHistogramStore.acceptFlushedCounts(countUpdates)
+        stateHistogramStore.acceptFlushedBytes(byteUpdates)
 
         return input
     }