+++ airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/dataflow/state/StateHistogram.kt
@@ -22,14 +22,14 @@ data class PartitionKey(
     val id: String,
 )
 
-open class Histogram<T>(open val map: ConcurrentMap<T, Long> = ConcurrentHashMap()) {
-    fun increment(key: T): Histogram<T> {
-        return this.apply { map.merge(key, 1, Long::plus) }
-    }
+open class Histogram<T>(private val map: ConcurrentMap<T, Long> = ConcurrentHashMap()) {
+    fun increment(key: T, quantity: Long): Histogram<T> =
+        this.apply { map.merge(key, quantity, Long::plus) }
 
-    fun merge(other: Histogram<T>): Histogram<T> {
-        return this.apply { other.map.forEach { map.merge(it.key, it.value, Long::plus) } }
-    }
+    fun merge(other: Histogram<T>): Histogram<T> =
+        this.apply { other.map.forEach { map.merge(it.key, it.value, Long::plus) } }
+
+    fun get(key: T): Long? = map[key]
 
     fun remove(key: T): Long? = map.remove(key)
 }