+++ airbyte-cdk/bulk/core/load/src/main/kotlin/io/airbyte/cdk/load/dataflow/state/StateStore.kt
@@ -5,53 +5,52 @@
 package io.airbyte.cdk.load.dataflow.state
 
 import io.airbyte.cdk.load.message.CheckpointMessage
+import io.github.oshai.kotlinlogging.KotlinLogging
 import jakarta.inject.Singleton
-import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.ConcurrentSkipListMap
+import java.util.concurrent.atomic.AtomicLong
 
 @Singleton
 class StateStore(
-    private val stateClient: StateKeyClient,
+    private val keyClient: StateKeyClient,
+    private val histogramStore: StateHistogramStore,
 ) {
-    // Counts of flushed messages by partition id
-    private val flushed: PartitionHistogram = PartitionHistogram(ConcurrentHashMap())
-    // Counts of expected messages by state id
-    private val expected: StateHistogram = StateHistogram(ConcurrentHashMap())
+    private val log = KotlinLogging.logger {}
+
     // state messages ordered by ID
     private val states = ConcurrentSkipListMap<StateKey, CheckpointMessage>()
+    // we publish states sequentially starting from 1
+    private val stateSequence = AtomicLong(1)
 
     fun accept(msg: CheckpointMessage) {
-        val key = stateClient.getStateKey(msg)
-        acceptExpectedCounts(key, msg.sourceStats!!.recordCount)
+        val key = keyClient.getStateKey(msg)
+        histogramStore.acceptExpectedCounts(key, msg.sourceStats!!.recordCount)
         states[key] = msg
     }
 
     fun remove(key: StateKey): CheckpointMessage = states.remove(key)!!
 
-    fun acceptFlushedCounts(value: PartitionHistogram): PartitionHistogram {
-        return flushed.merge(value)
-    }
-
-    fun acceptExpectedCounts(key: StateKey, count: Long): StateHistogram {
-        val inner = ConcurrentHashMap<StateKey, Long>()
-        inner[key] = count
-
-        return expected.merge(StateHistogram(inner))
-    }
-
     fun getNextComplete(): CheckpointMessage? {
         if (states.isEmpty()) return null
 
         val key = states.firstKey()
-        if (!isComplete(key)) return null
+        if (key.id != stateSequence.get()) return null
+        if (!histogramStore.isComplete(key)) return null
 
-        return states.remove(key)
+        stateSequence.incrementAndGet()
+        val msg = states.remove(key)
+        val stats = histogramStore.remove(key)
+
+        // Add count to stats (will always equal source stats)
+        // TODO: decide what we want to do with dest stats
+        msg!!.updateStats(
+            destinationStats = CheckpointMessage.Stats(stats.count),
+            totalRecords = stats.count,
+            totalBytes = stats.bytes,
+        )
+
+        return msg
     }
 
-    fun isComplete(key: StateKey): Boolean {
-        val expectedCount = expected.map[key]
-        val flushedCount = key.partitionKeys.sumOf { flushed.map[it] ?: 0 }
-
-        return expectedCount == flushedCount
-    }
+    fun hasStates(): Boolean = states.isNotEmpty()
 }