+++ airbyte-integrations/connectors/source-tiktok-marketing/unit_tests/integration/test_reports_hourly.py
@@ -207,7 +207,7 @@ class TestAdsReportHourly(TestCase):
             state=self.state(cursor=self.legacy_cursor),
         )
 
-        assert len(output.records) == 1
+        assert len(output.records) == 2
         assert output.state_messages[1].state.stream.stream_state.states == [
             {"cursor": {"stat_time_hour": self.legacy_cursor}, "partition": {"advertiser_id": self.advertiser_id, "parent_slice": {}}}
         ]
@@ -391,7 +391,7 @@ class TestAdGroupsReportsHourly(TestCase):
             state=self.state(),
         )
 
-        assert len(output.records) == 1
+        assert len(output.records) == 2
         assert output.state_messages[1].state.stream.stream_state.states == [
             {"cursor": {"stat_time_hour": self.cursor}, "partition": {"advertiser_id": self.advertiser_id, "parent_slice": {}}}
         ]
@@ -552,7 +552,7 @@ class TestAdvertisersReportsHourly(TestCase):
             state=self.state(),
         )
 
-        assert len(output.records) == 1
+        assert len(output.records) == 2
         assert output.state_messages[1].state.stream.stream_state.states == [
             {"cursor": {"stat_time_hour": self.cursor}, "partition": {"advertiser_id": self.advertiser_id, "parent_slice": {}}}
         ]
@@ -670,7 +670,7 @@ class TestCampaignsReportsHourly(TestCase):
             state=self.state(),
         )
 
-        assert len(output.records) == 1
+        assert len(output.records) == 2
         assert output.state_messages[1].state.stream.stream_state.states == [
             {"cursor": {"stat_time_hour": self.cursor}, "partition": {"advertiser_id": self.advertiser_id, "parent_slice": {}}}
         ]